Split Temporal activities into per-file modules
Committed-By-Agent: codex
