+++ apps/service/src/temporal/activities/setup.ts
@@ -1,0 +1,10 @@
+import { createRemoteEngine } from '@stripe/sync-engine'
+import type { PipelineConfig, SetupResult } from '@stripe/sync-engine'
+import type { ActivitiesContext } from './context.js'
+
+export function createSetupActivity(context: ActivitiesContext) {
+  return async function setup(config: PipelineConfig): Promise<SetupResult> {
+    const engine = createRemoteEngine(context.engineUrl, config)
+    return await engine.setup()
+  }
+}
