+++ apps/service/src/temporal/activities/teardown.ts
@@ -1,0 +1,10 @@
+import { createRemoteEngine } from '@stripe/sync-engine'
+import type { PipelineConfig } from '@stripe/sync-engine'
+import type { ActivitiesContext } from './context.js'
+
+export function createTeardownActivity(context: ActivitiesContext) {
+  return async function teardown(config: PipelineConfig): Promise<void> {
+    const engine = createRemoteEngine(context.engineUrl, config)
+    await engine.teardown()
+  }
+}
